+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ### Features
 
-* **maintenance:** add module introspection, cleanup, and install functions ([f291b6b](https://github.com/adnoctem/PSFoundation/commit/f291b6bc47e3fe5f72f83da47e68052b5da533c5))
+- **maintenance:** add module introspection, cleanup, and install functions ([f291b6b](https://github.com/adnoctem/PSFoundation/commit/f291b6bc47e3fe5f72f83da47e68052b5da533c5))